+++ CHANGELOG.md
@@ -80,6 +80,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed journal update after magic inventory is closed
 - fixed lever event not toggling the lever
 - fixed ConcurrentModificationException in PlayerData
+- fixed issue where the PacketInterceptor prints the message tag in the chat
 ### Security
 - fixed issue, where objectives that count things are out of sync with the database. This has also affected BungeeCord support